+++ demo/collision.js
@@ -21,9 +21,6 @@ const player = add([
 	rotate(0),
 	// area() component gives the object a collider, which enables collision checking
 	area(),
-// 	area({ shape: new Polygon([vec2(0), vec2(100), vec2(-100, 100)]) }),
-	area({ shape: new Rect(vec2(0), 12, 120) }),
-// 	area({ scale: 0.5 }),
 	// body() component makes an object respond to physics
 	body(),
 ])